Intermidiate commit.

+++ src/main/java/com/challengeteam/shop/service/TestDataService.java
@@ -2,7 +2,7 @@ package com.challengeteam.shop.service;
 
 public interface TestDataService {
 
-    void generateUsers(Long amount);
-    void generatePhones(Long amount);
+    void generateUsers(int amount);
+    void generatePhones(int amount);
 
 }
